+++ src/paradigma/__init__.py
@@ -3,6 +3,4 @@ from importlib.metadata import version
 
 __version__ = version("paradigma")
 
-from .imu_preprocessing import *
-
-__all__ = ["PreprocessingPipelineConfig"]
+__all__ = []
